third_party/WebKit/LayoutTests/http/tests/security/drag-drop-same-unique-origin-expected.txt

CONSOLE MESSAGE: line 22: PASS
